Build from github to fix tests in py37-dj20
py37-dj111 still fails for some reason. The only difference in the problematic method between 1.11.12 and 2.0 is whitespace. It's fine in py37-dj20 which is all we really care about so I'm inclined to ignore that issue. py37-dj111 isn't even tested on Travis CI so its failure won't be a problem.

Both django-celery-beat and adarnauth-esi have put out releases supporting dj20 so it's not necessary to build from their source.